requirements.txt
Flask
openai-whisper
gunicorn
ffmpeg-python
pymongo[srv]
python-dotenv